django/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.0.0
django-filter==24.3
djangorestframework==3.14.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
pytz==2025.1
sqlparse==0.5.2
